flw separate module
