+++ constants.go
@@ -5,11 +5,7 @@ import (
 	"fmt"
 )
 
-const (
-	protocolVersion = 0
-
-	DefaultPort = 2181
-)
+const protocolVersion = 0
 
 const (
 	opNotify          = 0
@@ -228,28 +224,3 @@ func (t EventType) String() string {
 	}
 	return "Unknown"
 }
-
-// Mode is used to build custom server modes (leader|follower|standalone).
-type Mode uint8
-
-func (m Mode) String() string {
-	if name := modeNames[m]; name != "" {
-		return name
-	}
-	return "unknown"
-}
-
-const (
-	ModeUnknown    Mode = iota
-	ModeLeader     Mode = iota
-	ModeFollower   Mode = iota
-	ModeStandalone Mode = iota
-)
-
-var (
-	modeNames = map[Mode]string{
-		ModeLeader:     "leader",
-		ModeFollower:   "follower",
-		ModeStandalone: "standalone",
-	}
-)
