+++ util.go
@@ -5,8 +5,6 @@ import (
 	"encoding/base64"
 	"fmt"
 	"math/rand"
-	"strconv"
-	"strings"
 	"unicode/utf8"
 )
 
@@ -32,21 +30,6 @@ func DigestACL(perms int32, user, password string) []ACL {
 	}
 	digest := base64.StdEncoding.EncodeToString(h.Sum(nil))
 	return []ACL{{perms, "digest", fmt.Sprintf("%s:%s", user, digest)}}
-}
-
-// FormatServers takes a slice of addresses, and makes sure they are in a format
-// that resembles <addr>:<port>. If the server has no port provided, the
-// DefaultPort constant is added to the end.
-func FormatServers(servers []string) []string {
-	srvs := make([]string, len(servers))
-	for i, addr := range servers {
-		if strings.Contains(addr, ":") {
-			srvs[i] = addr
-		} else {
-			srvs[i] = addr + ":" + strconv.Itoa(DefaultPort)
-		}
-	}
-	return srvs
 }
 
 // stringShuffle performs a Fisher-Yates shuffle on a slice of strings
